python-service/requirements.txt
fastapi
pandas
scikit-learn
uvicorn
google-genai
python-dotenv
openpyxl